refactor(controlplane): extract MCP supervisor and externalise its e2e test
Move runMCPSupervisor and its shared-key request-stamp helpers (newSharedKeyStamp,
signSharedKeyJWT) from server.go into a dedicated server_mcp.go. server.go drops
its dependencies on go-jose, go-jose/jwt and pkg/mcp/configapi.

Move server_mcp_configapi_test.go to package controlplane_test so its
test-only helpers (runServer, dialable) live outside the production package's
test scope. Export NewTestConfig (formerly newTestConfig in
server_mcp_late_enable_test.go) so the external configapi tests and the
internal server_test / late-enable tests can share one fixture builder; the
late-enable test still needs internal-package access for unexported handler
plumbing, so it stays in package controlplane.
